backend/requirements.txt
Flask==3.0.0
Flask-RESTful==0.3.10
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
PyJWT==2.8.0
yfinance>=1.0
beautifulsoup4==4.12.2
requests==2.31.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
colorama==0.4.6

